Use the beta compiler for building bootstrap tools
 ## Motivation

This avoids having to rebuild bootstrap and tidy each time you rebase
over master. In particular, it makes rebasing and running `x.py fmt` on
each commit in a branch significantly faster. It also avoids having to
rebuild bootstrap after setting `download-rustc = true`.

 ## Implementation

 Instead of extracting the CI artifacts directly to `stage0/`, extract
them to `master-rustc` instead. Continue to copy them to the proper
sysroots as necessary for all stages except stage 0.

 Note that since tools have to be built by the same compiler that built
`rustc-dev` and the standard library, the downloaded artifacts can't be
reused when building with the beta compiler. To make sure this is still
a good user experience, warn when building with the beta compiler, and
default to building with stage 2.